refactor: more jsonrpc into protocol dir
