+++ src/client/client.integration.test.ts
@@ -1,7 +1,7 @@
 import * as assert from 'assert'
-import { createMessageConnection, MessageConnection, MessageTransports } from '../jsonrpc2/connection'
-import { Trace } from '../jsonrpc2/trace'
 import { ClientCapabilities, InitializeParams, InitializeRequest, InitializeResult } from '../protocol'
+import { createMessageConnection, MessageConnection, MessageTransports } from '../protocol/jsonrpc2/connection'
+import { Trace } from '../protocol/jsonrpc2/trace'
 import { clientStateIsActive, getClientState } from '../test/helpers'
 import { createMessageTransports } from '../test/integration/helpers'
 import { Client, ClientState } from './client'
