+++ src/client/client.test.ts
@@ -1,7 +1,7 @@
 import * as assert from 'assert'
-import { createMessageConnection, MessageConnection, MessageTransports } from '../jsonrpc2/connection'
-import { MessageType as RPCMessageType } from '../jsonrpc2/messages'
 import { ClientCapabilities, RegistrationParams, UnregistrationParams } from '../protocol'
+import { createMessageConnection, MessageConnection, MessageTransports } from '../protocol/jsonrpc2/connection'
+import { MessageType as RPCMessageType } from '../protocol/jsonrpc2/messages'
 import { clientStateIs, getClientState } from '../test/helpers'
 import { createMessageTransports } from '../test/integration/helpers'
 import { Client, ClientOptions, ClientState } from './client'
