+++ src/extension/extensionHost.ts
@@ -1,8 +1,8 @@
 import { Subscription } from 'rxjs'
 import * as sourcegraph from 'sourcegraph'
-import { createMessageConnection, Logger, MessageConnection, MessageTransports } from '../jsonrpc2/connection'
-import { createWebWorkerMessageTransports } from '../jsonrpc2/transports/webWorker'
 import { InitializedNotification, InitializeParams, InitializeRequest, InitializeResult } from '../protocol'
+import { createMessageConnection, Logger, MessageConnection, MessageTransports } from '../protocol/jsonrpc2/connection'
+import { createWebWorkerMessageTransports } from '../protocol/jsonrpc2/transports/webWorker'
 import { createRegisterProviderFunctions } from './api/provider'
 import { Location } from './types/location'
 import { Position } from './types/position'
