+++ src/protocol/initialize.ts
@@ -1,6 +1,6 @@
-import { NotificationType, RequestType } from '../jsonrpc2/messages'
 import { ClientCapabilities } from './capabilities'
 import { ConfigurationCascade } from './configuration'
+import { NotificationType, RequestType } from './jsonrpc2/messages'
 
 /**
  * The initialize request is sent from the client to the server. It is sent once as the request after starting up
