+++ src/protocol/jsonrpc2/linkedMap.test.ts
