+++ src/protocol/jsonrpc2/linkedMap.ts
