+++ src/protocol/jsonrpc2/transports/stream.ts
